finalizacao das tasks

+++ DAG/ETL_scripts/remove_local.py
@@ -1,12 +1,13 @@
 import os
 import shutil
-from config import temp_json_files, temp_parquet_files
 
-json_parquet_paths = [temp_json_files, temp_parquet_files]
-
-def remove_local(paths=json_parquet_paths):
+def remove_local(temp_json_files, temp_parquet_files):
+	json_parquet_paths = [temp_json_files, temp_parquet_files]
 	for path in paths:
-		# remove pasta
-		shutil.rmtree(path, ignore_errors=True)
-		# refaz a pasta porem vazia
-		os.mkdir(path)
+		if os.path.isdir(path):
+			# remove pasta
+			shutil.rmtree(path, ignore_errors=True)
+			# refaz a pasta porem vazia
+			os.makedirs(path)
+		else:
+			print(f"Não foi possível remover os arquivos, pois {path} nao é uma pasta...")
